requirements.txt
Django==4.1.5
djongo==1.3.6
pymongo==3.12.3
djangorestframework==3.14.0
pytest==7.2.1
pytest-django==4.5.2
factory-boy==3.2.1